+++ lib/fittor.dart
@@ -1,6 +1,7 @@
 library;
 
 export 'src/fittor_code.dart';
+export 'src/fittor_custom_size.dart';
 export 'src/fittor_extensions.dart';
 export 'src/fittor_mixin.dart';
 export 'src/fittor_wrapper.dart';